+++ CHANGELOG.md
@@ -14,6 +14,7 @@ changes.
 
 - Add metadata url and hash to drep details [Issue 2911](https://github.com/IntersectMBO/govtool/issues/2911)
 - Add CC votes percentages, not voted and Ratification threshold
+- Add support for submitting all 7 governance action types [Issue 2258](https://github.com/IntersectMBO/govtool/issues/2258)
 
 ### Fixed
 
@@ -25,6 +26,7 @@ changes.
 
 - Change threshold visual representation in governance action votes
 - Resize governance action details columns
+- Update @intersect.mbo/pdf-ui to v0.6.0
 
 ### Removed